Moving source files to `src`
