--- manifest.json
@@ -1,64 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "Hacker News Companion",
-  "description": "Enhance Hacker News with smart navigation and AI summaries using local or cloud LLMs. Navigate easily with Vim-style shortcuts.",
-  "homepage_url": "https://github.com/levelup-apps/hn-enhancer",
-  "version": "1.0.8",
-  "icons": {
-    "16": "images/icon-16.png",
-    "32": "images/icon-32.png",
-    "48": "images/icon-48.png",
-    "128": "images/icon-128.png"
-  },
-  "browser_specific_settings": {
-    "gecko": {
-      "id": "addon@example.com"
-    },
-    "trial_tokens": [
-      "ArROXQIdYHn8CEwJCT02qMPHvAmS6ih8t1ZHiT5FpVVX3ayKlN47tmWAhqX/XnBiUQsuTd2hvPZUX4SevnGDPg8AAABzeyJvcmlnaW4iOiJjaHJvbWUtZXh0ZW5zaW9uOi8vbWJnZGJnZmRoZGhjbGJrYWNraGtsY2JlY2NtZWFrYmUiLCJmZWF0dXJlIjoiQUlTdW1tYXJpemF0aW9uQVBJIiwiZXhwaXJ5IjoxNzUzMTQyNDAwfQ==",
-      "AmvawA4HIg2FLhHRUdEaYpysamI5CI2wqhOkrSBLAe6gEwwNomjZYSeI/0Ni3gFZFQ6ED9BkQRwp31jRxRQbbwAAAABzeyJvcmlnaW4iOiJjaHJvbWUtZXh0ZW5zaW9uOi8va2hmY2FpbmVsY2FlZG1taGppY3BoYmtwaWdrbGVqZ2YiLCJmZWF0dXJlIjoiQUlTdW1tYXJpemF0aW9uQVBJIiwiZXhwaXJ5IjoxNzUzMTQyNDAwfQ=="
-    ]
-  },
-  "permissions": [
-    "storage"
-  ],
-  "host_permissions": [
-    "https://news.ycombinator.com/*",
-    "https://hn.algolia.com/*"
-  ],
-  "optional_host_permissions": [
-    "https://api.openai.com/v1/chat/completions/*",
-    "https://api.anthropic.com/v1/messages/*",
-    "http://localhost:11434/*",
-    "https://openrouter.ai/api/v1/*",
-    "https://api.deepseek.com/v1/chat/completions/*"
-  ],
-  "options_ui": {
-    "page": "options.html",
-    "open_in_tab": false
-  },
-  "action": {
-    "default_popup": "options.html"
-  },
-  "background": {
-    "service_worker": "background.js"
-  },
-  "content_scripts": [
-    {
-      "matches": [
-        "https://news.ycombinator.com/*"
-      ],
-      "css": [
-        "styles.css"
-      ],
-      "js": [
-        "summary-panel.js",
-        "content.js"
-      ]
-    }
-  ],
-  "web_accessible_resources": [{
-    "resources": ["page-script.js", "options-styles.css", "options.js"],
-    "matches": ["https://news.ycombinator.com/*"]
-  }]
-}
