+++ src/options.js
@@ -1,3 +1,4 @@
+// Save settings to Chrome storage
 async function saveSettings() {
     const providerSelection = document.querySelector('input[name="provider-selection"]:checked').id;
     const settings = {
@@ -25,6 +26,7 @@ async function saveSettings() {
 
     try {
         await chrome.storage.sync.set({ settings });
+        // Optional: Show save confirmation
         const saveButton = document.querySelector('button[type="submit"]');
         const originalText = saveButton.textContent;
         saveButton.textContent = 'Saved!';
@@ -57,6 +59,7 @@ async function sendBackgroundMessage(type, data) {
     return response.data;
 }
 
+// Fetch Ollama models from API
 async function fetchOllamaModels() {
     try {
         const data = await sendBackgroundMessage('FETCH_API_REQUEST', {
@@ -65,8 +68,10 @@ async function fetchOllamaModels() {
         });
 
         const selectElement = document.getElementById('ollama-model');
+        // Clear existing options
         selectElement.innerHTML = '';
 
+        // Add models to select element
         data.models.forEach(model => {
             const option = document.createElement('option');
             option.value = model.name;
@@ -74,6 +79,7 @@ async function fetchOllamaModels() {
             selectElement.appendChild(option);
         });
 
+        // If no models found, add a placeholder option
         if (data.models.length === 0) {
             const option = document.createElement('option');
             option.value = '';
@@ -82,6 +88,7 @@ async function fetchOllamaModels() {
         }
     } catch (error) {
         console.log('Error fetching Ollama models:', error);
+        // Handle error by adding an error option
         const selectElement = document.getElementById('ollama-model');
         selectElement.innerHTML = '';
         const option = document.createElement('option');
@@ -91,34 +98,41 @@ async function fetchOllamaModels() {
     }
 }
 
+// Load settings from Chrome storage
 async function loadSettings() {
     try {
         const data = await chrome.storage.sync.get('settings');
         const settings = data.settings;
 
         if (settings) {
+            // Set provider selection
             const providerRadio = document.getElementById(settings.providerSelection);
             if (providerRadio) providerRadio.checked = true;
 
+            // Set OpenAI settings
             if (settings.openai) {
                 document.getElementById('openai-key').value = settings.openai.apiKey || '';
                 document.getElementById('openai-model').value = settings.openai.model || 'gpt-4';
             }
 
+            // Set Anthropic settings
             if (settings.anthropic) {
                 document.getElementById('anthropic-key').value = settings.anthropic.apiKey || '';
                 document.getElementById('anthropic-model').value = settings.anthropic.model || 'claude-3-opus';
             }
 
+            // Set DeepSeek settings
             if (settings.deepseek) {
                 document.getElementById('deepseek-key').value = settings.deepseek.apiKey || '';
                 document.getElementById('deepseek-model').value = settings.deepseek.model || 'deepseek-chat';
             }
 
+            // Set Ollama settings
             if (settings.ollama) {
                 document.getElementById('ollama-model').value = settings.ollama.model || 'llama2';
             }
 
+            // Set OpenRouter settings
             if (settings.openrouter) {
                 document.getElementById('openrouter-key').value = settings.openrouter.apiKey || '';
                 document.getElementById('openrouter-model').value = settings.openrouter.model || 'anthropic/claude-3.5-sonnet';
@@ -129,24 +143,32 @@ async function loadSettings() {
     }
 }
 
+// Initialize event listeners and load settings
 document.addEventListener('DOMContentLoaded', async () => {
+    // Fetch Ollama models before loading other settings
     await fetchOllamaModels();
+
+    // Load saved settings
     await loadSettings();
 
+    // Add save button event listener
     const form = document.querySelector('form');
     form.addEventListener('submit', async (e) => {
         e.preventDefault();
         await saveSettings();
     });
 
+    // Add cancel button event listener
     const cancelButton = document.querySelector('button[type="button"]');
     cancelButton.addEventListener('click', () => {
         window.close();
     });
 
+    // Add radio button change listeners to enable/disable corresponding inputs
     const radioButtons = document.querySelectorAll('input[name="provider-selection"]');
     radioButtons.forEach(radio => {
         radio.addEventListener('change', () => {
+            // Enable/disable input fields based on selection
             const openaiInputs = document.querySelectorAll('#openai-key, #openai-model');
             const anthropicInputs = document.querySelectorAll('#anthropic-key, #anthropic-model');
             const deepseekInputs = document.querySelectorAll('#deepseek-key, #deepseek-model');
@@ -161,6 +183,7 @@ document.addEventListener('DOMContentLoaded', async () => {
         });
     });
 
+    // Initial trigger of radio button change event to set initial state
     const checkedRadio = document.querySelector('input[name="provider-selection"]:checked');
     if (checkedRadio) {
         checkedRadio.dispatchEvent(new Event('change'));
