+++ src/summary-panel.js
