+++ vite.config.js
@@ -1,19 +1,25 @@
-import { defineConfig } from 'vite';
+import { defineConfig } from "vite";
 
-export default defineConfig({
-    build: {
-        rollupOptions: {
-            input: 'src/content.js', // Entry point
-            output: [
-                {
-                    format: 'iife', // Immediately Invoked Function Expression
-                    file: 'dist/chrome/content.bundle.js', // Output file for Chrome
-                },
-                {
-                    format: 'iife', // Immediately Invoked Function Expression
-                    file: 'dist/firefox/content.bundle.js', // Output file for Firefox
-                }
-            ],
+export default defineConfig(({ mode }) => {
+    return {
+        build: {
+            outDir: "dist", // Output directory
+            minify: mode === 'production' ? true : false, // Minify only in production mode
+            rollupOptions: {
+                input: "src/content.js", // Entry point
+                output: [
+                    {
+                        format: "iife", // Immediately Invoked Function Expression
+                        entryFileNames: "chrome/content.bundle.js", // Output file for Chrome
+                        dir: "dist", // Output directory
+                    },
+                    {
+                        format: "iife", // Immediately Invoked Function Expression
+                        entryFileNames: "firefox/content.bundle.js", // Output file for Firefox
+                        dir: "dist", // Output directory
+                    },
+                ],
+            },
         },
-    },
+    };
 });
